fix(proxy): move transport-health marker into the support domain
The proxy architecture matrix allows http_bridge and websocket to import
only their listed domains (support among them, streaming not), so the
shared transport-failure marker belongs in _service/support.py rather
than a new streaming module. Also align the test harness override of
_handle_websocket_connect_error with the base signature for ty.
